requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.2.1
Django==5.2.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
greenlet==3.2.2
inflection==0.5.1
iniconfig==2.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
playwright==1.52.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pyee==13.0.0
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
